Page template added
