+++ src/components/footer/footer.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
+import {Link} from 'react-router-dom';
 
 export const Footer = () => {
   return (
     <footer className="footer container">
-      <a className="footer__logo-link" href="/">
-        <img className="footer__logo" src="/img/logo.svg" alt="6 cities logo" width="64" height="33" />
-      </a>
+      <Link to="/" className="footer__logo-link">
+        <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
+      </Link>
     </footer>
   );
 };
